crawler/requirements.txt
requests>=2.31.0
selenium>=4.16.0
urllib3>=2.1.0
python-magic>=0.4.27
beautifulsoup4>=4.12.0
pillow>=10.4.0,<12.0.0
openpyxl>=3.1.2
pypdf>=3.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-chroma>=0.1.0
chromadb>=0.4.0
tiktoken>=0.5.0
tqdm>=4.65.0
pandas>=2.0.0 